requirements.txt
Django==4.1.7
django-lifecycle==1.0.0
django-widget-tweaks==1.4.12
python-dateutil==2.8.2